+++ Cargo.toml
@@ -13,4 +13,4 @@ exclude = ["proofs", "delta-inspect"]
 split-debuginfo = "unpacked"
 
 [patch.crates-io]
-object_store = { git = "https://github.com/apache/arrow-rs", rev = "b34adcce427c6cb74fde2d99bc95b8731b7ceda7" }
+object_store = { git = "https://github.com/apache/arrow-rs", rev = "86446ea88b6fc3e11b47aeedb9eb24f4d69b2f13" }
